+++ pkg/anonymity/queue/queue.go
@@ -54,7 +54,7 @@ func NewQBProblemProcessor(pSettings ISettings, pClient client.IClient) IQBProbl
 		fClient:   pClient,
 		fMainPool: &sMainPool{
 			fQueue:     make(chan net_message.IMessage, queuePoolCap[0]*consumersCap),
-			fConsumers: make(map[string]uint64, 64),
+			fConsumers: make(map[string]uint64, 128),
 			fRawQueue: func() map[uint64]chan []byte {
 				m := make(map[uint64]chan []byte, consumersCap)
 				for i := uint64(0); i < consumersCap; i++ {
@@ -119,13 +119,13 @@ func (p *sQBProblemProcessor) runMainPoolFiller(pCtx context.Context, pCancel fu
 		pWG.Done()
 		pCancel()
 	}()
-	for i := uint64(0); ; i++ {
+	for i := uint64(0); ; i = (i + 1) % p.fSettings.GetConsumersCap() {
 		select {
 		case <-pCtx.Done():
 			return
 		case <-time.After(p.fSettings.GetQueuePeriod()):
 			break // next consumer
-		case msg := <-p.fMainPool.fRawQueue[i%p.fSettings.GetConsumersCap()]:
+		case msg := <-p.fMainPool.fRawQueue[i]:
 			if err := p.pushMessage(pCtx, p.fMainPool.fQueue, msg); err != nil {
 				return
 			}
@@ -150,7 +150,7 @@ func (p *sQBProblemProcessor) EnqueueMessage(pPubKey asymmetric.IPubKey, pBytes 
 	hash := pPubKey.GetHasher().ToString()
 	v, ok := p.fMainPool.fConsumers[hash]
 	if !ok {
-		v = uint64(len(p.fMainPool.fConsumers)+1) % p.fSettings.GetConsumersCap()
+		v = uint64(len(p.fMainPool.fConsumers)) % p.fSettings.GetConsumersCap()
 		p.fMainPool.fConsumers[hash] = v
 	}
 	p.fMainPool.fMutex.Unlock()