+++ pkg/anonymity/types.go
@@ -4,21 +4,18 @@ import (
 	"context"
 	"time"
 
+	"github.com/number571/go-peer/pkg/anonymity/adapters"
+	"github.com/number571/go-peer/pkg/anonymity/queue"
 	"github.com/number571/go-peer/pkg/crypto/asymmetric"
 	"github.com/number571/go-peer/pkg/logger"
-	"github.com/number571/go-peer/pkg/network"
-	"github.com/number571/go-peer/pkg/network/anonymity/queue"
 	"github.com/number571/go-peer/pkg/payload"
 	"github.com/number571/go-peer/pkg/storage/database"
 	"github.com/number571/go-peer/pkg/types"
 )
 
-type IHandlerF func(
-	context.Context,
-	INode,
-	asymmetric.IPubKey,
-	[]byte,
-) ([]byte, error)
+type (
+	IHandlerF func(context.Context, INode, asymmetric.IPubKey, []byte) ([]byte, error)
+)
 
 type INode interface {
 	types.IRunner
@@ -26,10 +23,10 @@ type INode interface {
 
 	GetLogger() logger.ILogger
 	GetSettings() ISettings
+	GetAdapter() adapters.IAdapter
 	GetKVDatabase() database.IKVDatabase
-	GetNetworkNode() network.INode
-	GetMessageQueue() queue.IQBProblemProcessor
 	GetMapPubKeys() asymmetric.IMapPubKeys
+	GetQBProcessor() queue.IQBProblemProcessor
 
 	SendPayload(context.Context, asymmetric.IPubKey, payload.IPayload64) error
 	FetchPayload(context.Context, asymmetric.IPubKey, payload.IPayload32) ([]byte, error)